Authentication and Authorization
- Added Passport authentication using HTTP for initial login and JWT for future requests to API
- Updated all endpoints with JWT authentication

+++ index.js
@@ -8,6 +8,10 @@ const express = require('express'),
 const app = express();
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
+    
+let auth = require('./auth')(app);
+const passport = require('passport');
+require('./passport');
 
 const Movies = Models.Movie;
 const Users = Models.User;
@@ -22,8 +26,8 @@ app.get('/', (req, res) => {
     res.send('Welcome to myFlix!')
 });
 
-// returns JSON list of movies
-app.get('/movies', async (req, res) => {
+// READ - returns JSON list of all movies
+app.get('/movies', passport.authenticate('jwt', { session: false }), async (req, res) => {
     await Movies.find()
     .then((movies) => {
         res.status(201).json(movies);
@@ -35,7 +39,7 @@ app.get('/movies', async (req, res) => {
 });
 
 // READ - returns JSON object of individual movie
-app.get('/movies/:Title', async (req, res) => {
+app.get('/movies/:Title', passport.authenticate('jwt', { session: false }), async (req, res) => {
     await Movies.findOne({ Title: req.params.Title })
     .then((movie) => {
         res.json(movie);
@@ -46,7 +50,8 @@ app.get('/movies/:Title', async (req, res) => {
     });
 });
 
-app.get('/movies/genre/:genreName', async (req, res) => {
+// READ - returns JSON object of individual genre and description
+app.get('/movies/genre/:genreName', passport.authenticate('jwt', { session: false }), async (req, res) => {
     await Movies.findOne({ 'Genre.Name': req.params.genreName })
     .then((genre) => {
         res.json(genre.Genre);
@@ -57,7 +62,8 @@ app.get('/movies/genre/:genreName', async (req, res) => {
     });
 });
 
-app.get('/movies/director/:directorName', async (req, res) => {
+// READ - returns a JSON object of individual director data(name, bio, birth year, death year)
+app.get('/movies/director/:directorName', passport.authenticate('jwt', { session: false }), async (req, res) => {
     await Movies.findOne({ 'Director.Name': req.params.directorName})
     .then((data) => {
         res.json(data.Director);
@@ -68,16 +74,7 @@ app.get('/movies/director/:directorName', async (req, res) => {
     });
 });
 
-// Add a user
-/* Expected JSON format
-{
-ID: Integer,
-Username: String,
-Password: String,
-Email: String,
-Birthday: Date
-} */
-
+// CREATE - add new user
 app.post('/users', async (req, res) => {
     await Users.findOne({ Username: req.body.Username })
         .then((user) => {
@@ -104,7 +101,11 @@ app.post('/users', async (req, res) => {
         });
 });
 
-app.put('/users/:Username', async (req, res) => {
+// UPDATE - change user information by username
+app.put('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    if(req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }    
     await Users.findOneAndUpdate({ Username: req.params.Username },
     {$set:
         {
@@ -124,7 +125,11 @@ app.put('/users/:Username', async (req, res) => {
     });
 });
 
-app.post('/users/:Username/movies/:MovieID', async (req, res) => {
+// CREATE - add new movie to FavoriteMovies by Username ane MovieID
+app.post('/users/:Username/movies/:MovieID', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    if(req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }  
     await Users.findOneAndUpdate(
         { Username: req.params.Username },
         {$push: { FavoriteMovies: req.params.MovieID }},
@@ -139,7 +144,11 @@ app.post('/users/:Username/movies/:MovieID', async (req, res) => {
     });
 });
 
-app.delete('/users/:Username/movies/:MovieID', async (req, res) => {
+// Delete - delete movie from FavoriteMovies by Username ane MovieID
+app.delete('/users/:Username/movies/:MovieID', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    if(req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }  
     await Users.findOneAndUpdate(
         { Username: req.params.Username },
         {$pull: { FavoriteMovies: req.params.MovieID }},
@@ -154,7 +163,11 @@ app.delete('/users/:Username/movies/:MovieID', async (req, res) => {
     });
 });
 
-app.delete('/users/:Username', async (req, res) => {
+// DELETE - delete user from database
+app.delete('/users/:Username', passport.authenticate('jwt', { session: false }), async (req, res) => {
+    if(req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    } 
     await Users.findOneAndDelete({ Username: req.params.Username })
     .then((user) => {
         if(!user) {
